BWS-223: + refetch after deleting with saving of the current page to objects table
