+++ frontend/src/components/eam/components/groups-table/groups-table.tsx
@@ -50,6 +50,7 @@ const GroupsTable: FC<Props> = ({ onGroupDelete }) => {
     countItems,
     onLoad: handleLoad,
     from: Pagination.INITIAL_FROM_COUNT,
+    entities: groups.length,
   });
 
   const handleGroupsReload = (): void => {
